requirements.txt
asgiref==3.4.1
Django==3.2.4
django-cors-headers==3.9.0
django-filter==21.1
djangorestframework==3.12.4
gunicorn==20.1.0
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0
